Remove: ADCI1: setData(T data) method
The signature of setData(T data) method of ADCI1 interface
is ambiguous for recognizing whether
the caller of the method want to handle the data container
as a scalar or as a vector(array), from the code of caller-side.
Therefore, it has been removed by this commit.

After this commit, the caller should call
setData(T data, int offset) for setting a scalar value, or
setData(T data, int[] lengths) for setting a vector value,
instead of calling setData(T data) method.