Remove: PACI1: methods for editing permission values from other plug-ins
Following methods have been removed:

* setPermissionValue
* getPermissionValue
* storePermissionValues
* restorePermissionValues

Their specifications are "draft" yet.
It is completely unclear
whether we will support them in future.

Hence, in future,
they might be going to be "deprecated" methods,
if we will determine that
they should not be supported forever.
So we should be careful of making decisions
about specifications of the above methods.

On the other hand, we want finalize
the specification of PACI1 as soon as possible.
Therefore, we have removed them from PACI1.
We will soon define PACI2 (as a draft)
and moved the above methods to it.